backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
django-cors-headers==3.14.0
django-extensions==3.2.3
